+++ example/lib/universal_ui/real_ui.dart
@@ -1,4 +1,4 @@
-import 'dart:ui' as ui;
+import 'dart:ui' if (dart.library.html) 'dart:ui_web' as ui;
 
 class PlatformViewRegistry {
   static void registerViewFactory(String viewId, dynamic cb) {