+++ example/lib/widgets/demo_scaffold.dart
@@ -1,5 +1,5 @@
 import 'dart:convert';
-import 'dart:io';
+import 'dart:io' show Platform;
 
 import 'package:filesystem_picker/filesystem_picker.dart';
 import 'package:flutter/foundation.dart';