+++ flutter_quill_extensions/lib/embeds/embed_types.dart
@@ -1,4 +1,4 @@
-import 'dart:io';
+import 'dart:io' show File;
 import 'dart:typed_data';
 
 import 'package:flutter/material.dart';