+++ flutter_quill_extensions/lib/embeds/toolbar/image_video_utils.dart
@@ -1,4 +1,4 @@
-import 'dart:io';
+import 'dart:io' show File;
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';