+++ flutter_quill_extensions/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.5.1
+- Fix warrning "The platformViewRegistry getter is deprecated and will be removed in a future release. Please import it from dart:ui_web instead."
+- Add QuillImageUtilities class
+- Small improvemenets
+- Allow to use the mobile context menu on desktop by force using it
+
 ## 0.5.0
 - Migrated from `gallery_saver` to `gal` for saving images
 - Added callbacks for greater control of editing images